+++ Src/Virtual Printer Solution/VirtualPrinter.HostedService.HttpSystem/Hosted Services/HttpListenerService.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  *  This file is part of Virtual ZPL Printer.
  *
  *  Virtual ZPL Printer is free software: you can redistribute it and/or modify
@@ -14,7 +14,6 @@
  *  You should have received a copy of the GNU General Public License
  *  along with Virtual ZPL Printer.  If not, see <https://www.gnu.org/licenses/>.
  */
-using System.Collections.Specialized;
 using System.Net;
 using System.Text;
 using Diamond.Core.Extensions.Hosting;
@@ -33,7 +32,7 @@ namespace VirtualPrinter.HostedService.HttpSystem
 	// Implements the Zebra printer HTTP web server endpoints:
 	//   GET /printer/dir  — lists stored .ZPL format files
 	//   GET /printer/zpl  — returns the content of a specific stored format
-	public class HttpListenerService : HostedServiceTemplate
+	public partial class HttpListenerService : HostedServiceTemplate
 	{
 		public const int DefaultHttpPort = 9200;
 
@@ -196,7 +195,7 @@ namespace VirtualPrinter.HostedService.HttpSystem
 				{
 					await this.HandleZplMetaSaveAsync(context);
 				}
-								else if (path.Equals("/printer/grf", StringComparison.OrdinalIgnoreCase))
+				else if (path.Equals("/printer/grf", StringComparison.OrdinalIgnoreCase))
 				{
 					await this.HandleGrfAsync(context);
 				}
@@ -229,689 +228,11 @@ namespace VirtualPrinter.HostedService.HttpSystem
 			}
 		}
 
-		private async Task HandlePrinterIndexAsync(HttpListenerContext context)
-		{
-			// Pick up optional ?status=ok&file=X or ?status=err&msg=X set by the upload redirect.
-			NameValueCollection qs    = context.Request.QueryString;
-			string statusParam        = qs["status"] ?? string.Empty;
-			string statusFile         = qs["file"]   ?? string.Empty;
-			string statusMsg          = qs["msg"]    ?? string.Empty;
-
-			var sb = new StringBuilder();
-			sb.AppendLine("<!DOCTYPE html><html><head><meta charset='utf-8'>");
-			sb.AppendLine("<title>Virtual ZPL Printer — Flash Memory</title>");
-			sb.AppendLine("<style>");
-			sb.AppendLine("  body { font-family: monospace; margin: 2em; }");
-			sb.AppendLine("  h2 { border-bottom: 1px solid #ccc; padding-bottom: 4px; }");
-			sb.AppendLine("  table { border-collapse: collapse; width: 100%; margin-bottom: 1em; }");
-			sb.AppendLine("  th, td { text-align: left; padding: 4px 12px; border-bottom: 1px solid #eee; }");
-			sb.AppendLine("  th { background: #f4f4f4; }");
-			sb.AppendLine("  .none { color: #999; font-style: italic; }");
-			sb.AppendLine("  .upload { margin-bottom: 2em; display: flex; align-items: center; gap: 8px; }");
-			sb.AppendLine("  .upload input[type=file] { font-family: monospace; }");
-			sb.AppendLine("  .upload button { padding: 3px 10px; cursor: pointer; }");
-			sb.AppendLine("  .msg { padding: 6px 12px; border-radius: 4px; margin-bottom: 1em; }");
-			sb.AppendLine("  .msg.ok  { background: #d4edda; color: #155724; }");
-			sb.AppendLine("  .msg.err { background: #f8d7da; color: #721c24; }");
-			sb.AppendLine("</style></head><body>");
-			sb.AppendLine("<h1>Virtual ZPL Printer — Flash Memory</h1>");
-
-			if (statusParam.Equals("ok", StringComparison.OrdinalIgnoreCase))
-				sb.AppendLine($"<div class='msg ok'>Uploaded <strong>{System.Net.WebUtility.HtmlEncode(statusFile)}</strong> successfully.</div>");
-			else if (statusParam.Equals("err", StringComparison.OrdinalIgnoreCase))
-				sb.AppendLine($"<div class='msg err'>Upload failed: {System.Net.WebUtility.HtmlEncode(statusMsg)}</div>");
-
-			// --- Single upload form (auto-detects ^DF vs ~DG) ---
-			sb.AppendLine("<form class='upload' method='post' action='/printer/upload' enctype='multipart/form-data'>");
-			sb.AppendLine("  <input type='file' name='file' accept='.zpl,.pl'>");
-			sb.AppendLine("  <button type='submit'>Upload</button>");
-			sb.AppendLine("  <small>Files with ^DF are stored as Formats; files with ~DG are stored as Graphics; files with neither are imported as a Format using the filename.</small>");
-			sb.AppendLine("</form>");
-			// --- Formats (ZPL) ---
-			sb.AppendLine("<h2>Formats (ZPL)</h2>");
-
-			DirectoryInfo fmtDir = this.ZplFormatService.FormatDirectory;
-			FileInfo[] fmtFiles  = fmtDir.Exists ? fmtDir.GetFiles() : [];
-
-			if (fmtFiles.Length == 0)
-			{
-				sb.AppendLine("<p class='none'>No formats stored. Send a ^DF job to the printer to store one.</p>");
-			}
-			else
-			{
-				sb.AppendLine("<table><tr><th>Name</th><th>Size</th><th>Modified</th><th></th></tr>");
-
-				foreach (FileInfo file in fmtFiles.OrderBy(f => f.Name))
-				{
-					if (file.Name.EndsWith(".meta.json", StringComparison.OrdinalIgnoreCase)) continue;
-					int underscore = file.Name.IndexOf('_');
-					if (underscore <= 0) continue;
-
-					string dev  = file.Name[..underscore];
-					string rest = file.Name[(underscore + 1)..];
-					int dot     = rest.LastIndexOf('.');
-					if (dot <= 0) continue;
-
-					string oname       = rest[..dot];
-					string otype       = rest[(dot + 1)..];
-					string displayName = $"{dev}:{rest}";
-					string viewHref    = $"/printer/zpl?dev={dev}&oname={oname}&otype={otype}";
-
-					string deleteForm = $"<form method='post' action='/printer/delete' style='display:inline'><input type='hidden' name='key' value='{file.Name}'><button type='submit'>Delete</button></form>";
-					sb.AppendLine($"<tr><td>{displayName}</td><td>{file.Length:N0} B</td><td>{file.LastWriteTime:yyyy-MM-dd HH:mm:ss}</td><td><a href='{viewHref}'>View</a>&nbsp;&nbsp;{deleteForm}</td></tr>");
-				}
-
-				sb.AppendLine("</table>");
-			}
-
-			// --- Graphics (GRF) ---
-			sb.AppendLine("<h2>Graphics (GRF)</h2>");
-
-			DirectoryInfo grfDir = this.GrfStorageService.GrfDirectory;
-			FileInfo[] grfFiles  = grfDir.Exists ? grfDir.GetFiles() : [];
-
-			if (grfFiles.Length == 0)
-			{
-				sb.AppendLine("<p class='none'>No graphics stored. Send a ~DG job to the printer to store one.</p>");
-			}
-			else
-			{
-				sb.AppendLine("<table><tr><th>Name</th><th>Size</th><th>Modified</th><th></th></tr>");
-
-				foreach (FileInfo file in grfFiles.OrderBy(f => f.Name))
-				{
-					int underscore = file.Name.IndexOf('_');
-					if (underscore <= 0) continue;
-
-					string dev         = file.Name[..underscore];
-					string rest        = file.Name[(underscore + 1)..];
-					string displayName = $"{dev}:{rest}";
-					string viewHref    = $"/printer/grf?dev={dev}&filename={rest}";
-
-					string deleteForm = $"<form method='post' action='/printer/delete' style='display:inline'><input type='hidden' name='key' value='{file.Name}'><button type='submit'>Delete</button></form>";
-					sb.AppendLine($"<tr><td>{displayName}</td><td>{file.Length:N0} B</td><td>{file.LastWriteTime:yyyy-MM-dd HH:mm:ss}</td><td><a href='{viewHref}'>View</a>&nbsp;&nbsp;{deleteForm}</td></tr>");
-				}
-
-				sb.AppendLine("</table>");
-			}
-
-			sb.AppendLine("</body></html>");
-
-			await WriteResponseAsync(context, "text/html", sb.ToString());
-
-			this.Logger.LogDebug("Responded to /printer index.");
-		}
-
-		private async Task HandleDirAsync(HttpListenerContext context)
-		{
-			// Build an HTML page listing stored .ZPL format files.
-
-			var sb = new StringBuilder();
-			sb.AppendLine("<html><body>");
-			sb.AppendLine("<h2>E: DIRECTORY</h2><ul>");
-
-			DirectoryInfo dir = this.ZplFormatService.FormatDirectory;
-
-			if (dir.Exists)
-			{
-				foreach (FileInfo file in dir.GetFiles())
-				{
-					string storedName = file.Name;
-
-					int underscore = storedName.IndexOf('_');
-					if (underscore <= 0) continue;
-
-					string dev  = storedName[..underscore];
-					string rest = storedName[(underscore + 1)..];
-					int dot     = rest.LastIndexOf('.');
-					if (dot <= 0) continue;
-
-					string oname       = rest[..dot];
-					string otype       = rest[(dot + 1)..];
-					string displayName = $"{dev}:{rest}";
-
-					sb.AppendLine($"  <li><a href=\"zpl?dev={dev}&oname={oname}&otype={otype}\">{displayName}</a></li>");
-				}
-			}
-
-			sb.AppendLine("</ul></body></html>");
-
-			await WriteResponseAsync(context, "text/html", sb.ToString());
-
-			this.Logger.LogInformation("Responded to /printer/dir.");
-		}
-
-		private async Task HandleZplAsync(HttpListenerContext context)
-	{
-		NameValueCollection qs = context.Request.QueryString;
-		string dev   = qs["dev"]   ?? string.Empty;
-		string oname = qs["oname"] ?? string.Empty;
-		string otype = qs["otype"] ?? string.Empty;
-
-		if (string.IsNullOrEmpty(dev) || string.IsNullOrEmpty(oname) || string.IsNullOrEmpty(otype))
-		{
-			context.Response.StatusCode = 400;
-			context.Response.Close();
-			return;
-		}
-
-		string key      = $"{dev}_{oname}.{otype}";
-		string filePath = Path.Combine(this.ZplFormatService.FormatDirectory.FullName, key);
-
-		if (!File.Exists(filePath))
-		{
-			this.Logger.LogWarning("HTTP /printer/zpl: format '{key}' not found.", key);
-			context.Response.StatusCode = 404;
-			context.Response.Close();
-			return;
-		}
-
-		string fileContent = await File.ReadAllTextAsync(filePath);
-		string displayName = $"{dev}:{oname}.{otype}";
-		string statusParam = qs["status"] ?? string.Empty;
-
-		// Load saved ^FN preview values from meta sidecar file.
-		string metaPath = Path.Combine(this.ZplFormatService.FormatDirectory.FullName, $"{key}.meta.json");
-		string metaJson = File.Exists(metaPath) ? await File.ReadAllTextAsync(metaPath) : "{}";
-
-		// Serialize dynamic values as JS literals to avoid injection.
-		string jsdev         = System.Text.Json.JsonSerializer.Serialize(dev);
-		string jsoname       = System.Text.Json.JsonSerializer.Serialize(oname);
-		string jsotype       = System.Text.Json.JsonSerializer.Serialize(otype);
-		string jsContent     = System.Text.Json.JsonSerializer.Serialize(fileContent);
-		string jsDisplayName = System.Net.WebUtility.HtmlEncode(displayName);
-
-		var sb = new StringBuilder();
-		sb.AppendLine("<!DOCTYPE html><html><head><meta charset='utf-8'>");
-		sb.AppendLine($"<title>Edit ZPL Script -- {jsDisplayName}</title>");
-		sb.AppendLine("<style>");
-		sb.AppendLine("  body { font-family: monospace; margin: 2em; }");
-		sb.AppendLine("  h2 { border-bottom: 1px solid #ccc; padding-bottom: 4px; }");
-		sb.AppendLine("  textarea { width: 100%; height: 320px; font-family: monospace; font-size: 13px; padding: 8px; box-sizing: border-box; border: 1px solid #ccc; resize: vertical; }");
-		sb.AppendLine("  .btn-row { margin-top: 8px; display: flex; gap: 8px; align-items: center; }");
-		sb.AppendLine("  .btn-row button { padding: 4px 14px; cursor: pointer; }");
-		sb.AppendLine("  .fn-section { margin-top: 12px; }");
-		sb.AppendLine("  .fn-section summary { cursor: pointer; color: #555; font-size: 13px; user-select: none; }");
-		sb.AppendLine("  .fn-table { border-collapse: collapse; margin-top: 6px; }");
-		sb.AppendLine("  .fn-table th { text-align: left; padding: 3px 12px 3px 0; color: #555; font-size: 12px; border-bottom: 1px solid #ddd; }");
-		sb.AppendLine("  .fn-table td { padding: 3px 12px 3px 0; }");
-		sb.AppendLine("  .fn-table td:first-child { color: #888; font-size: 12px; min-width: 55px; }");
-		sb.AppendLine("  .fn-table input { font-family: monospace; font-size: 13px; padding: 2px 6px; border: 1px solid #ccc; width: 320px; }");
-		sb.AppendLine("  .preview-area { margin-top: 1.5em; }");
-		sb.AppendLine("  .preview-area img { border: 1px solid #ccc; background: #fff; max-width: 100%; display: block; }");
-		sb.AppendLine("  .placeholder { color: #999; font-style: italic; margin-top: 8px; }");
-		sb.AppendLine("  .msg { padding: 6px 12px; border-radius: 4px; margin-bottom: 1em; }");
-		sb.AppendLine("  .msg.ok  { background: #d4edda; color: #155724; }");
-		sb.AppendLine("  .msg.err { background: #f8d7da; color: #721c24; }");
-		sb.AppendLine("  a { color: #00c; }");
-		sb.AppendLine("</style></head><body>");
-		sb.AppendLine("<h2>Edit ZPL Script</h2>");
-		sb.AppendLine($"<p><strong>{jsDisplayName}</strong> &nbsp; <a href='/printer'>&#8592; Directory</a></p>");
-
-		if (statusParam.Equals("ok", StringComparison.OrdinalIgnoreCase))
-			sb.AppendLine("<div class='msg ok'>Saved successfully.</div>");
-		else if (statusParam.Equals("err", StringComparison.OrdinalIgnoreCase))
-			sb.AppendLine($"<div class='msg err'>Save failed: {System.Net.WebUtility.HtmlEncode(qs["msg"] ?? string.Empty)}</div>");
-
-		sb.AppendLine($"<textarea id='zpl'>{System.Net.WebUtility.HtmlEncode(fileContent)}</textarea>");
-
-		sb.AppendLine("<details class='fn-section' id='fn-details'>");
-		sb.AppendLine("  <summary id='fn-summary'>Field Values (^FN) -- loading...</summary>");
-		sb.AppendLine("  <table class='fn-table'>");
-		sb.AppendLine("    <thead><tr><th>Field</th><th>Preview Value</th></tr></thead>");
-		sb.AppendLine("    <tbody id='fn-body'></tbody>");
-		sb.AppendLine("  </table>");
-		sb.AppendLine("</details>");
-
-		sb.AppendLine("<div class='btn-row'>");
-		sb.AppendLine("  <button onclick='previewLabel()'>Preview Label</button>");
-		sb.AppendLine("  <button onclick='saveLabel()'>Save</button>");
-		sb.AppendLine("  <button onclick='resetLabel()'>Reset</button>");
-		sb.AppendLine("</div>");
-		sb.AppendLine("<div class='preview-area'>");
-		sb.AppendLine("  <img id='preview-img' style='display:none' alt='Label preview'>");
-		sb.AppendLine("  <p class='placeholder' id='preview-ph'>Click Preview Label to render via Labelary.</p>");
-		sb.AppendLine("</div>");
-		sb.AppendLine("<script>");
-		sb.AppendLine("var dev = " + jsdev + ";");
-		sb.AppendLine("var oname = " + jsoname + ";");
-		sb.AppendLine("var otype = " + jsotype + ";");
-		sb.AppendLine("var originalContent = " + jsContent + ";");
-		sb.AppendLine("var savedMeta = " + metaJson + ";");
-		sb.AppendLine("function resetLabel() {");
-		sb.AppendLine("  document.getElementById('zpl').value = originalContent;");
-		sb.AppendLine("  buildFnTable();");
-		sb.AppendLine("}");
-		sb.AppendLine("function parseFnFields(zpl) {");
-		sb.AppendLine("  var nums = []; var seen = {};");
-		sb.AppendLine("  var re = /\\^FN(\\d+)/gi; var m;");
-		sb.AppendLine("  while ((m = re.exec(zpl)) !== null) {");
-		sb.AppendLine("    var n = parseInt(m[1], 10);");
-		sb.AppendLine("    if (!seen[n]) { seen[n] = true; nums.push(n); }");
-		sb.AppendLine("  }");
-		sb.AppendLine("  nums.sort(function(a,b){return a-b;}); return nums;");
-		sb.AppendLine("}");
-		sb.AppendLine("function getFnValues() {");
-		sb.AppendLine("  var vals = {};");
-		sb.AppendLine("  document.querySelectorAll('#fn-body tr').forEach(function(tr) {");
-		sb.AppendLine("    var num = tr.getAttribute('data-fn');");
-		sb.AppendLine("    var inp = tr.querySelector('input');");
-		sb.AppendLine("    if (num && inp) vals[num] = inp.value;");
-		sb.AppendLine("  });");
-		sb.AppendLine("  return vals;");
-		sb.AppendLine("}");
-		sb.AppendLine("function buildFnTable() {");
-		sb.AppendLine("  var zpl = document.getElementById('zpl').value;");
-		sb.AppendLine("  var nums = parseFnFields(zpl);");
-		sb.AppendLine("  var tbody = document.getElementById('fn-body');");
-		sb.AppendLine("  var summary = document.getElementById('fn-summary');");
-		sb.AppendLine("  var details = document.getElementById('fn-details');");
-		sb.AppendLine("  tbody.innerHTML = '';");
-		sb.AppendLine("  if (nums.length === 0) {");
-		sb.AppendLine("    summary.textContent = 'No ^FN fields found'; details.open = false; return;");
-		sb.AppendLine("  }");
-		sb.AppendLine("  summary.textContent = 'Field Values (^FN) -- ' + nums.length + ' field' + (nums.length > 1 ? 's' : '');");
-		sb.AppendLine("  details.open = true;");
-		sb.AppendLine("  nums.forEach(function(n) {");
-		sb.AppendLine("    var tr = document.createElement('tr');");
-		sb.AppendLine("    tr.setAttribute('data-fn', String(n));");
-		sb.AppendLine("    var td1 = document.createElement('td'); td1.textContent = '^FN' + n;");
-		sb.AppendLine("    var td2 = document.createElement('td');");
-		sb.AppendLine("    var inp = document.createElement('input');");
-		sb.AppendLine("    inp.type = 'text'; inp.value = savedMeta[String(n)] || '';");
-		sb.AppendLine("    inp.placeholder = '(leave blank to skip)';");
-		sb.AppendLine("    inp.addEventListener('input', scheduleSaveMeta);");
-		sb.AppendLine("    td2.appendChild(inp); tr.appendChild(td1); tr.appendChild(td2); tbody.appendChild(tr);");
-		sb.AppendLine("  });");
-		sb.AppendLine("}");
-		sb.AppendLine("function injectFnValues(zpl, vals) {");
-		sb.AppendLine("  return zpl.replace(/\\^FN(\\d+)\\^FS/gi, function(m, n) {");
-		sb.AppendLine("    var v = vals[n]; return v ? '^FN' + n + '^FV' + v + '^FS' : m;");
-		sb.AppendLine("  });");
-		sb.AppendLine("}");
-		sb.AppendLine("var metaSaveTimer = null;");
-		sb.AppendLine("function scheduleSaveMeta() { clearTimeout(metaSaveTimer); metaSaveTimer = setTimeout(saveMeta, 800); }");
-		sb.AppendLine("function saveMeta() {");
-		sb.AppendLine("  var vals = getFnValues(); savedMeta = vals;");
-		sb.AppendLine("  var p = 'dev=' + encodeURIComponent(dev) + '&oname=' + encodeURIComponent(oname)");
-		sb.AppendLine("        + '&otype=' + encodeURIComponent(otype) + '&meta=' + encodeURIComponent(JSON.stringify(vals));");
-		sb.AppendLine("  fetch('/printer/zpl/meta', { method: 'POST', headers: { 'Content-Type': 'application/x-www-form-urlencoded' }, body: p });");
-		sb.AppendLine("}");
-		sb.AppendLine("async function previewLabel() {");
-		sb.AppendLine("  saveMeta();");
-		sb.AppendLine("  var zpl = document.getElementById('zpl').value;");
-		sb.AppendLine("  var ph = document.getElementById('preview-ph');");
-		sb.AppendLine("  var img = document.getElementById('preview-img');");
-		sb.AppendLine("  var injected = injectFnValues(zpl, getFnValues());");
-		sb.AppendLine("  ph.textContent = 'Rendering...'; img.style.display = 'none';");
-		sb.AppendLine("  try {");
-		sb.AppendLine("    var resp = await fetch('/printer/preview', { method: 'POST', headers: { 'Content-Type': 'text/plain' }, body: injected });");
-		sb.AppendLine("    if (!resp.ok) { ph.textContent = 'Preview failed: ' + await resp.text(); return; }");
-		sb.AppendLine("    img.src = URL.createObjectURL(await resp.blob()); img.style.display = 'block'; ph.textContent = '';");
-		sb.AppendLine("  } catch(e) { ph.textContent = 'Preview error: ' + e; }");
-		sb.AppendLine("}");
-		sb.AppendLine("async function saveLabel() {");
-		sb.AppendLine("  var p = 'dev=' + encodeURIComponent(dev) + '&oname=' + encodeURIComponent(oname)");
-		sb.AppendLine("        + '&otype=' + encodeURIComponent(otype) + '&content=' + encodeURIComponent(document.getElementById('zpl').value);");
-		sb.AppendLine("  var resp = await fetch('/printer/zpl', { method: 'POST', headers: { 'Content-Type': 'application/x-www-form-urlencoded' }, body: p, redirect: 'follow' });");
-		sb.AppendLine("  window.location.href = resp.url || window.location.href;");
-		sb.AppendLine("}");
-		sb.AppendLine("function refreshFnTable() {");
-		sb.AppendLine("  var cur = getFnValues();");
-		sb.AppendLine("  Object.keys(cur).forEach(function(k) { if (cur[k]) savedMeta[k] = cur[k]; });");
-		sb.AppendLine("  buildFnTable();");
-		sb.AppendLine("}");
-		sb.AppendLine("document.getElementById('zpl').addEventListener('blur', refreshFnTable);");
-		sb.AppendLine("buildFnTable();");
-		sb.AppendLine("</script>");
-		sb.AppendLine("</body></html>");
-
-		await WriteResponseAsync(context, "text/html", sb.ToString());
-		this.Logger.LogInformation("Responded to /printer/zpl edit page for '{dev}:{oname}.{otype}'.", dev, oname, otype);
-	}
-
-	private async Task HandleZplSaveAsync(HttpListenerContext context)
-	{
-		using MemoryStream ms = new();
-		await context.Request.InputStream.CopyToAsync(ms);
-		string body = Encoding.UTF8.GetString(ms.ToArray());
-
-		string dev = null, oname = null, otype = null, content = null;
-
-		foreach (string pair in body.Split('&'))
-		{
-			int eq = pair.IndexOf('=');
-			if (eq < 0) continue;
-			string name  = Uri.UnescapeDataString(pair[..eq].Replace('+', ' '));
-			string value = Uri.UnescapeDataString(pair[(eq + 1)..].Replace('+', ' '));
-			switch (name.ToLowerInvariant())
-			{
-				case "dev":     dev     = value; break;
-				case "oname":   oname   = value; break;
-				case "otype":   otype   = value; break;
-				case "content": content = value; break;
-			}
-		}
-
-		if (string.IsNullOrEmpty(dev) || string.IsNullOrEmpty(oname) || string.IsNullOrEmpty(otype) || content == null)
-		{
-			this.Redirect(context, $"/printer/zpl?dev={dev}&oname={oname}&otype={otype}&status=err&msg=Missing+parameters");
-			return;
-		}
-
-		string key      = $"{dev}_{oname}.{otype}";
-		string filePath = Path.Combine(this.ZplFormatService.FormatDirectory.FullName, key);
-
-		if (!File.Exists(filePath))
-		{
-			this.Redirect(context, $"/printer/zpl?dev={dev}&oname={oname}&otype={otype}&status=err&msg=File+not+found");
-			return;
-		}
-
-		await File.WriteAllTextAsync(filePath, content);
-		this.Logger.LogInformation("Saved ZPL edit for '{dev}:{oname}.{otype}'.", dev, oname, otype);
-		this.Redirect(context, $"/printer/zpl?dev={dev}&oname={oname}&otype={otype}&status=ok");
-	}
-
-	private async Task HandlePreviewAsync(HttpListenerContext context)
-	{
-		if (this.LabelConfiguration == null)
-		{
-			context.Response.StatusCode = 503;
-			await WriteResponseAsync(context, "text/plain", "No printer running. Start a virtual printer first.");
-			return;
-		}
-
-		using MemoryStream ms = new();
-		await context.Request.InputStream.CopyToAsync(ms);
-		string zplContent = Encoding.UTF8.GetString(ms.ToArray()).Trim();
-
-		if (!zplContent.StartsWith("^XA", StringComparison.OrdinalIgnoreCase))
-			zplContent = $"^XA\r\n{zplContent}\r\n^XZ";
-
-		IGetLabelResponse response = await this.LabelService.GetLabelAsync(this.LabelConfiguration, zplContent);
-
-		if (!response.Result || response.Label == null || response.Label.Length == 0)
-		{
-			string error = response.Error ?? "Labelary returned no image.";
-			this.Logger.LogWarning("HTTP /printer/preview: Labelary render failed: {error}", error);
-			context.Response.StatusCode = 502;
-			await WriteResponseAsync(context, "text/plain", $"Render failed: {error}");
-			return;
-		}
-
-		context.Response.ContentType     = "image/png";
-		context.Response.ContentLength64 = response.Label.Length;
-		context.Response.StatusCode      = 200;
-		await context.Response.OutputStream.WriteAsync(response.Label);
-		context.Response.Close();
-		this.Logger.LogInformation("Rendered label preview via Labelary ({bytes} bytes).", response.Label.Length);
-	}
-
-	private async Task HandleZplMetaSaveAsync(HttpListenerContext context)
-	{
-		using MemoryStream ms = new();
-		await context.Request.InputStream.CopyToAsync(ms);
-		string body = Encoding.UTF8.GetString(ms.ToArray());
-
-		string dev = null, oname = null, otype = null, meta = null;
-
-		foreach (string pair in body.Split('&'))
-		{
-			int eq = pair.IndexOf('=');
-			if (eq < 0) continue;
-			string name  = Uri.UnescapeDataString(pair[..eq].Replace('+', ' '));
-			string value = Uri.UnescapeDataString(pair[(eq + 1)..].Replace('+', ' '));
-			switch (name.ToLowerInvariant())
-			{
-				case "dev":   dev   = value; break;
-				case "oname": oname = value; break;
-				case "otype": otype = value; break;
-				case "meta":  meta  = value; break;
-			}
-		}
-
-		if (string.IsNullOrEmpty(dev) || string.IsNullOrEmpty(oname) || string.IsNullOrEmpty(otype) || meta == null)
-		{
-			context.Response.StatusCode = 400;
-			context.Response.Close();
-			return;
-		}
-
-		string key      = $"{dev}_{oname}.{otype}";
-		string metaPath = Path.Combine(this.ZplFormatService.FormatDirectory.FullName, $"{key}.meta.json");
-		await File.WriteAllTextAsync(metaPath, meta);
-
-		context.Response.StatusCode = 204;
-		context.Response.Close();
-		this.Logger.LogDebug("Saved FN meta for '{key}'.", key);
-	}
-	private async Task HandleUploadAsync(HttpListenerContext context)
-		{
-			// Parse multipart/form-data, extract the file content as ZPL text, then
-			// run it through the same storage pipeline as a live ZPL print job.
-			// Both ~DG (GRF) and ^DF (format template) are captured automatically.
-
-			string contentType = context.Request.ContentType ?? string.Empty;
-
-			if (!contentType.Contains("multipart/form-data", StringComparison.OrdinalIgnoreCase))
-			{
-				this.Redirect(context, "/printer?status=err&msg=Expected+multipart/form-data");
-				return;
-			}
-
-			// Extract the boundary token from Content-Type.
-			string boundary = null;
-
-			foreach (string part in contentType.Split(';'))
-			{
-				string trimmed = part.Trim();
-
-				if (trimmed.StartsWith("boundary=", StringComparison.OrdinalIgnoreCase))
-				{
-					boundary = trimmed["boundary=".Length..].Trim('"');
-					break;
-				}
-			}
-
-			if (string.IsNullOrEmpty(boundary))
-			{
-				this.Redirect(context, "/printer?status=err&msg=Missing+multipart+boundary");
-				return;
-			}
-
-			// Read the full request body.
-			using MemoryStream ms = new();
-			await context.Request.InputStream.CopyToAsync(ms);
-			string body = Encoding.UTF8.GetString(ms.ToArray());
-
-			// Locate the file part: skip boundary + part headers, read until closing boundary.
-			string partStart  = $"--{boundary}\r\n";
-			string partEnd    = $"\r\n--{boundary}";
-			int headerEnd     = body.IndexOf("\r\n\r\n", StringComparison.Ordinal);
-			int contentStart  = headerEnd >= 0 ? headerEnd + 4 : 0;
-			int contentEnd    = body.IndexOf(partEnd, contentStart, StringComparison.Ordinal);
-			string zplContent = contentEnd >= 0
-				? body[contentStart..contentEnd]
-				: body[contentStart..];
-
-			if (string.IsNullOrWhiteSpace(zplContent))
-			{
-				this.Redirect(context, "/printer?status=err&msg=Uploaded+file+was+empty");
-				return;
-			}
-
-			// Extract the original filename from Content-Disposition for the success message.
-			string filename = "file";
-
-			foreach (string line in body[..contentStart].Split('\n'))
-			{
-				if (!line.TrimStart().StartsWith("Content-Disposition", StringComparison.OrdinalIgnoreCase)) continue;
-
-				foreach (string token in line.Split(';'))
-				{
-					string t = token.Trim();
-
-					if (t.StartsWith("filename=", StringComparison.OrdinalIgnoreCase))
-					{
-						filename = t["filename=".Length..].Trim('"', '\'', '\r', '\n');
-						break;
-					}
-				}
-			}
-
-			bool hasDf = zplContent.Contains("^DF", StringComparison.OrdinalIgnoreCase);
-			bool hasDg = zplContent.Contains("~DG", StringComparison.OrdinalIgnoreCase);
-
-			if (!hasDf && !hasDg)
-			{
-				// No explicit ^DF or ~DG — treat the whole file as a format template and
-				// wrap it automatically using the uploaded filename as the format name.
-				string baseName = Path.GetFileNameWithoutExtension(filename);
-				string ext      = Path.GetExtension(filename).TrimStart('.').ToUpperInvariant();
-				if (string.IsNullOrEmpty(ext)) ext = "ZPL";
-				string dfName   = $"E:{baseName}.{ext}";
-
-				// Strip the outer ^XA / ^XZ so the inner content can be re-wrapped cleanly.
-				string inner = zplContent;
-				int xaIdx    = inner.IndexOf("^XA", StringComparison.OrdinalIgnoreCase);
-				if (xaIdx >= 0) inner = inner[(xaIdx + 3)..];
-				int xzIdx    = inner.LastIndexOf("^XZ", StringComparison.OrdinalIgnoreCase);
-				if (xzIdx >= 0) inner = inner[..xzIdx];
-
-				zplContent = $"^XA\r\n^DF{dfName}^FS\r\n{inner.Trim()}\r\n^XZ";
-
-				this.Logger.LogInformation("No ^DF found in '{filename}' — auto-wrapped as {dfName}.", filename, dfName);
-			}
-
-			// Run through the standard flash-storage pipeline — same as a live ZPL job.
-			await this.GrfStorageService.SaveGrfFromZplAsync(zplContent);
-			await this.ZplFormatService.SaveFormatFromZplAsync(zplContent);
-
-			this.Logger.LogInformation("Uploaded '{filename}' via HTTP.", filename);
-
-			string encodedFile = Uri.EscapeDataString(filename);
-			this.Redirect(context, $"/printer?status=ok&file={encodedFile}");
-		}
-
 		private void Redirect(HttpListenerContext context, string location)
 		{
 			context.Response.StatusCode        = 302;
 			context.Response.RedirectLocation  = location;
 			context.Response.Close();
-		}
-
-		private async Task HandleDeleteAsync(HttpListenerContext context)
-		{
-			// Read application/x-www-form-urlencoded body to get the storage key (e.g. "E_FILENAME.ZPL").
-			using MemoryStream ms = new();
-			await context.Request.InputStream.CopyToAsync(ms);
-			string body = Encoding.UTF8.GetString(ms.ToArray());
-
-			string key = null;
-
-			foreach (string pair in body.Split('&'))
-			{
-				int eq = pair.IndexOf('=');
-				if (eq < 0) continue;
-
-				string name  = Uri.UnescapeDataString(pair[..eq].Replace('+', ' '));
-				string value = Uri.UnescapeDataString(pair[(eq + 1)..].Replace('+', ' '));
-
-				if (name.Equals("key", StringComparison.OrdinalIgnoreCase))
-				{
-					key = value;
-					break;
-				}
-			}
-
-			if (string.IsNullOrEmpty(key))
-			{
-				this.Redirect(context, "/printer?status=err&msg=Missing+key");
-				return;
-			}
-
-			// Try Formats directory first, then Graphics.
-			string[] candidates =
-			[
-				Path.Combine(this.ZplFormatService.FormatDirectory.FullName, key),
-				Path.Combine(this.GrfStorageService.GrfDirectory.FullName,   key),
-			];
-
-			string matched = candidates.FirstOrDefault(File.Exists);
-
-			if (matched == null)
-			{
-				this.Redirect(context, $"/printer?status=err&msg=File+not+found");
-				return;
-			}
-
-			File.Delete(matched);
-			this.Logger.LogInformation("Deleted flash file '{key}'.", key);
-
-			string encodedKey = Uri.EscapeDataString(key);
-			this.Redirect(context, $"/printer?status=ok&file={encodedKey}+deleted");
-		}
-
-		private async Task HandleGrfAsync(HttpListenerContext context)
-		{
-			// Renders a stored GRF image by constructing minimal ZPL that inlines the
-			// ~DG blob and references it with ^XG, then passes it to Labelary.
-			//
-			// Query string: ?dev=E&filename=IMAGE.GRF
-
-			NameValueCollection qs = context.Request.QueryString;
-			string dev      = qs["dev"]      ?? string.Empty;
-			string filename = qs["filename"] ?? string.Empty;
-
-			if (string.IsNullOrEmpty(dev) || string.IsNullOrEmpty(filename))
-			{
-				context.Response.StatusCode = 400;
-				context.Response.Close();
-				return;
-			}
-
-			if (this.LabelConfiguration == null)
-			{
-				await WriteResponseAsync(context, "text/plain", "No printer is running. Start a virtual printer first.");
-				return;
-			}
-
-			string key      = $"{dev}_{filename}";
-			string filePath = Path.Combine(this.GrfStorageService.GrfDirectory.FullName, key);
-
-			if (!File.Exists(filePath))
-			{
-				this.Logger.LogWarning("HTTP /printer/grf: GRF '{key}' not found.", key);
-				context.Response.StatusCode = 404;
-				context.Response.Close();
-				return;
-			}
-
-			string grfBlob = await File.ReadAllTextAsync(filePath);
-
-			// Construct minimal ZPL: inline the ~DG blob then recall it with ^XG.
-			// ^FO0,0 positions the image at the top-left corner of the label.
-			string zpl = $"^XA\r\n{grfBlob.Trim()}\r\n^FO0,0^XG{dev}:{filename}^FS\r\n^XZ";
-
-			IGetLabelResponse response = await this.LabelService.GetLabelAsync(this.LabelConfiguration, zpl);
-
-			if (!response.Result || response.Label == null || response.Label.Length == 0)
-			{
-				string error = response.Error ?? "Labelary returned no image.";
-				this.Logger.LogWarning("HTTP /printer/grf: Labelary render failed for '{key}': {error}", key, error);
-				await WriteResponseAsync(context, "text/plain", $"Render failed: {error}");
-				return;
-			}
-
-			context.Response.ContentType     = "image/png";
-			context.Response.ContentLength64 = response.Label.Length;
-			context.Response.StatusCode      = 200;
-
-			await context.Response.OutputStream.WriteAsync(response.Label);
-			context.Response.Close();
-
-			this.Logger.LogInformation("Rendered GRF '{dev}:{filename}' via Labelary.", dev, filename);
 		}
 
 		private static async Task WriteResponseAsync(HttpListenerContext context, string contentType, string body)
